renaming and moving api functions
