+++ client/src/components/Auth/LoginForm.js
@@ -11,7 +11,7 @@ import {
   InputGroupText,
   InputGroup,
 } from 'reactstrap';
-import { login } from 'api/authFunctions';
+import { login } from 'api/userFunctions';
 import { GlobalDispatchContext } from 'components/GlobalState';
 
 const LoginForm = (prop) => {
