+++ client/src/views/MyFiles.js
@@ -17,7 +17,7 @@
 */
 import { useEffect, useState } from 'react';
 
-import { getFiles } from 'api/authFunctions';
+import { getFiles } from 'api/fileFunctions';
 import Header from 'components/Headers/Header.js';
 import FileTable from 'components/File/FileTable.js';
