+++ client/src/views/Starred.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { getFiles } from 'api/authFunctions';
+import { getFiles } from 'api/fileFunctions';
 import Header from 'components/Headers/Header.js';
 import FileTable from 'components/File/FileTable.js';
